Lots of Updates
Waves and wave infos were converted to Scriptable Objects + Money System + Money gain by hitting enemies and money spending by upgrades. + Range Circle

+++ Assets/Scripts/UpgradeData.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class UpgradeData : ScriptableObject
 {
     public float fireRate;
-    public float firePower;
+    public int firePower;
     public float range;
-    public float upgradeCost;
+    public int upgradeCost;
 
 }
